tools/flowsaver/pyproject.toml
[project]
name = "flowloganalysis"
version = "0.1.0"
description = "Parse and analyze Azure VNet flow logs with DuckDB"
requires-python = ">=3.10"
dependencies = [
    "duckdb>=1.0",
    "pytz",
]

[project.optional-dependencies]
azure = [
    "azure-storage-blob>=12.0",
    "azure-identity>=1.0",
]
dev = [
    "pytest>=7.0",
]

[project.scripts]
flowlog = "flowloganalysis.cli:main"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"
